Implement callbacks for transport actions
Use webkit_web_view_run_javascript() to execute JavaScript code in the web
process, making it actually possible to cause the buttons to perform actions
in the Groove Basin web UI.

Unfortunately, this depends on "_debug_player" being set in the global
scope, which is an implementation detail. At some point it would be better
to have a minimal client for the Groove Basin client protocol, and use that
for the transport buttons, and to get song-change notifications.

This implementation should be enough for the moment.